core: move modifiers to separate files
