+++ Blasphemous.LostDreams/LostDreams.cs
@@ -11,8 +11,8 @@ using Blasphemous.LostDreams.Items.Penitences;
 using Blasphemous.LostDreams.Items.QuestItems;
 using Blasphemous.LostDreams.Items.RosaryBeads;
 using Blasphemous.LostDreams.Items.SwordHearts;
-using Blasphemous.LostDreams.Levels;
 using Blasphemous.LostDreams.Levels.Loaders;
+using Blasphemous.LostDreams.Levels.Modifiers;
 using Blasphemous.LostDreams.Npc;
 using Blasphemous.ModdingAPI;
 using UnityEngine;
